api/requirements.txt
attrs==19.1.0
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
datapunt-config-loader==1.1.0
datapunt-objectstore==2018.12.10
debtcollector==1.21.0
defusedxml==0.5.0
Django==2.1.8
django-choices==1.6.2
django-cors-headers==2.5.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-csv==2.1.0
djangorestframework-gis==0.14
djangorestframework-xml==1.4.0
drf-amsterdam==0.3.1
drf-extensions==0.4.0
drf-yasg==1.15.0
entrypoints==0.3
filelock==3.0.10
flake8==3.7.7
idna==2.8
inflection==0.3.1
iso8601==0.1.12
itypes==1.1.0
Jinja2==2.10
jsonschema==3.0.1
keystoneauth1==3.13.1
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==1.6.0
msgpack==0.6.1
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.6.0
oslo.config==6.8.1
oslo.i18n==3.23.1
oslo.serialization==2.28.2
oslo.utils==3.40.3
pbr==5.1.3
pluggy==0.9.0
psycopg2-binary==2.7.7
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.3.1
pyrsistent==0.14.11
python-keystoneclient==3.19.0
python-swiftclient==3.7.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
rfc3986==1.2.0
ruamel.yaml==0.15.89
sentry-sdk==0.7.9
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
toml==0.10.0
tox==3.8.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.18
virtualenv==16.4.3
wrapt==1.11.1
xlrd==1.2.0
